feat: middleware implementation done

+++ routes/api.php
@@ -20,15 +20,13 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-
 Route::post('/login', [UserController::class, 'login']);
 
-Route::group(['middleware' => ['auth:sanctum']], function() {
+Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::post('/register', [UserController::class, 'register'])->middleware('isadmin', 'auth');
-    Route::get('/users', [UserController::class, 'getUsers'])->name('users');
-    Route::get('/userprofile', [UserController::class, 'userProfile']);
+    Route::get('/users', [UserController::class, 'getUsers'])->name('users')->middleware('isadmin', 'auth');
+    Route::get('/userprofile', [UserController::class, 'userProfile'])->middleware('isadmin', 'auth');
     Route::get('/logout', [UserController::class, 'logout']);
-    Route::patch('/update/{id}', [UserController::class, 'update'])->name('update');
-    Route::delete('/delete/{id}', [UserController::class, 'delete']);
-
+    Route::patch('/update/{id}', [UserController::class, 'update'])->name('update')->middleware('isadmin', 'auth');
+    Route::delete('/delete/{id}', [UserController::class, 'delete'])->middleware('isadmin', 'auth');
 });
